Modularize / Move common configuration and create a minimal OGC API Records app.
