+++ src/apps/geonetwork/src/main/java/org/geonetwork/config/WebSecurityConfiguration.java
@@ -10,25 +10,17 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import org.apache.commons.lang3.StringUtils;
-import org.geonetwork.domain.repository.UserRepository;
 import org.geonetwork.proxy.HttpProxyPolicyAgentAuthorizationManager;
-import org.geonetwork.security.DatabaseUserAuthProperties;
-import org.geonetwork.security.DatabaseUserDetailsService;
-import org.geonetwork.security.GeoNetworkUserService;
+import org.geonetwork.security.GeoNetworkOAuth2UserService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.crypto.password.StandardPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.access.expression.DefaultWebSecurityExpressionHandler;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 @Configuration
@@ -68,12 +60,7 @@ public class WebSecurityConfiguration {
                 .logout(logout -> logout.logoutRequestMatcher(new AntPathRequestMatcher("/api/user/signout"))
                         .logoutSuccessHandler((request, response, authentication) -> {
                             handleRedirectParam(request, response, homeUrl);
-                        }))
-                .csrf(csrf -> csrf.disable());
-
-        //    http.sessionManagement(
-        //        session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-
+                        }));
         return http.build();
     }
 
@@ -81,36 +68,5 @@ public class WebSecurityConfiguration {
             throws IOException {
         String redirectUrl = request.getParameter("redirectUrl");
         response.sendRedirect(StringUtils.isNotEmpty(redirectUrl) ? redirectUrl : request.getContextPath() + homeUrl);
-    }
-
-    @Bean
-    @SuppressWarnings("deprecation")
-    public PasswordEncoder passwordEncoder(@Value("${geonetwork.security.passwordSalt}") String salt) {
-        return new StandardPasswordEncoder(salt);
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService(
-            @Value("${geonetwork.security.checkUsernameOrEmail: 'USERNAME_OR_EMAIL'}")
-                    DatabaseUserAuthProperties checkUsernameOrEmail,
-            PasswordEncoder passwordEncoder,
-            UserRepository userRepository,
-            GeoNetworkUserService geoNetworkUserService) {
-        return new DatabaseUserDetailsService(
-                checkUsernameOrEmail, passwordEncoder, geoNetworkUserService, userRepository);
-    }
-
-    @Bean
-    public RoleHierarchyImpl roleHierarchy() {
-        final RoleHierarchyImpl roleHierarchy = RoleHierarchyImpl.fromHierarchy(
-                "ROLE_Administrator > ROLE_UserAdmin\n ROLE_UserAdmin > ROLE_Reviewer\n ROLE_Reviewer > ROLE_Editor\n ROLE_Editor > ROLE_RegisteredUser\n ROLE_RegisteredUser > ROLE_Guest");
-        return roleHierarchy;
-    }
-
-    @Bean
-    public DefaultWebSecurityExpressionHandler expressionHandler() {
-        DefaultWebSecurityExpressionHandler expressionHandler = new DefaultWebSecurityExpressionHandler();
-        expressionHandler.setRoleHierarchy(roleHierarchy());
-        return expressionHandler;
     }
 }
