+++ src/apps/ogc-api-records/pom.xml
@@ -14,9 +14,9 @@
     <version>5.0.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>gn-main-app</artifactId>
-  <name>GeoNetwork main app</name>
-  <description>The main GeoNetwork application</description>
+  <artifactId>gn-ogcapi-records-app</artifactId>
+  <name>GeoNetwork OGC API Records app</name>
+  <description></description>
   <dependencies>
     <dependency>
       <groupId>org.geonetwork</groupId>
@@ -30,22 +30,7 @@
     </dependency>
     <dependency>
       <groupId>org.geonetwork</groupId>
-      <artifactId>gn-indexing</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.geonetwork</groupId>
       <artifactId>gn-configuration</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.geonetwork</groupId>
-      <artifactId>gn-proxy</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.geonetwork</groupId>
-      <artifactId>gn-searching</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -55,17 +40,7 @@
     </dependency>
     <dependency>
       <groupId>org.geonetwork</groupId>
-      <artifactId>gn-editing</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.geonetwork</groupId>
       <artifactId>gn-security</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.geonetwork</groupId>
-      <artifactId>gn-4</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
