+++ src/modules/ogcapi-records/ogcapi-records-implementation/src/main/java/org/geonetwork/ogcapi/configuration/OgcApiRecordOpenApiConfiguration.java
@@ -4,47 +4,19 @@
  * available at the root application directory.
  */
 
-package org.geonetwork.config;
+package org.geonetwork.ogcapi.configuration;
 
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Contact;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.info.License;
-import io.swagger.v3.oas.models.media.*;
+import io.swagger.v3.oas.models.media.ArraySchema;
 import org.geonetwork.ogcapi.service.indexConvert.dynamic.DynamicPropertiesFacade;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-/** OpenAPI configuration. */
 @Configuration
-public class OpenApiConfiguration {
+public class OgcApiRecordOpenApiConfiguration {
 
     @Autowired
     DynamicPropertiesFacade dynamicPropertiesFacade;
-
-    //    @Value("${geonetwork.openapi.url:'http://localhost:7979'}")
-    //    private String serverUrl;
-
-    /** OpenAPI configuration. */
-    @Bean
-    public OpenAPI myOpenApi() {
-        Contact contact = new Contact();
-        contact.setEmail("geonetwork-users@lists.sourceforge.net");
-        contact.setName("GeoNetwork opensource");
-        contact.setUrl("https://geonetwork-opensource.org/");
-
-        License license = new License().name("GPL 2.0").url("https://www.gnu.org/licenses/old-licenses/lgpl-2.0.html");
-
-        Info info = new Info()
-                .title("GeoNetwork API")
-                .version("5.0.0")
-                .contact(contact)
-                .description("This API exposes endpoints to GeoNetwork API.")
-                .license(license);
-
-        return new OpenAPI().info(info);
-    }
 
     /**
      * adds in the dynamic properties
