+++ src/shared/configuration/src/main/java/org/geonetwork/configuration/SecurityConfiguration.java
@@ -1,0 +1,54 @@
+/*
+ * (c) 2003 Open Source Geospatial Foundation - all rights reserved
+ * This code is licensed under the GPL 2.0 license,
+ * available at the root application directory.
+ */
+package org.geonetwork.configuration;
+
+import org.geonetwork.domain.repository.UserRepository;
+import org.geonetwork.security.DatabaseUserAuthProperties;
+import org.geonetwork.security.DatabaseUserDetailsService;
+import org.geonetwork.security.GeoNetworkUserService;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.crypto.password.StandardPasswordEncoder;
+import org.springframework.security.web.access.expression.DefaultWebSecurityExpressionHandler;
+
+@Configuration
+public class SecurityConfiguration {
+
+    @Bean
+    @SuppressWarnings("deprecation")
+    public PasswordEncoder passwordEncoder(@Value("${geonetwork.security.passwordSalt}") String salt) {
+        return new StandardPasswordEncoder(salt);
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService(
+            @Value("${geonetwork.security.checkUsernameOrEmail: 'USERNAME_OR_EMAIL'}")
+                    DatabaseUserAuthProperties checkUsernameOrEmail,
+            PasswordEncoder passwordEncoder,
+            UserRepository userRepository,
+            GeoNetworkUserService geoNetworkUserService) {
+        return new DatabaseUserDetailsService(
+                checkUsernameOrEmail, passwordEncoder, geoNetworkUserService, userRepository);
+    }
+
+    @Bean
+    public RoleHierarchyImpl roleHierarchy() {
+        final RoleHierarchyImpl roleHierarchy = RoleHierarchyImpl.fromHierarchy(
+                "ROLE_Administrator > ROLE_UserAdmin\n ROLE_UserAdmin > ROLE_Reviewer\n ROLE_Reviewer > ROLE_Editor\n ROLE_Editor > ROLE_RegisteredUser\n ROLE_RegisteredUser > ROLE_Guest");
+        return roleHierarchy;
+    }
+
+    @Bean
+    public DefaultWebSecurityExpressionHandler expressionHandler() {
+        DefaultWebSecurityExpressionHandler expressionHandler = new DefaultWebSecurityExpressionHandler();
+        expressionHandler.setRoleHierarchy(roleHierarchy());
+        return expressionHandler;
+    }
+}
