+++ src/shared/pom.xml
@@ -22,6 +22,7 @@
 
   <modules>
     <module>testing</module>
+    <module>configuration</module>
     <module>constants</module>
     <module>domain</module>
     <module>utility</module>
