+++ src/shared/security/src/main/java/org/geonetwork/security/GeoNetworkSsoConfiguration.java
@@ -3,7 +3,7 @@
  * This code is licensed under the GPL 2.0 license,
  * available at the root application directory.
  */
-package org.geonetwork.config;
+package org.geonetwork.security;
 
 import java.util.HashMap;
 import java.util.Map;
